feat(tx_builder): Generalize `TxOrdering` inputs and outputs
`TxOrdering` is made generic by exposing the generic from
`TxSort` function. The benefit is that we're no longer limited
to ordering lists of only `TxIn` or `TxOut`.

We use bitcoin `TxIn` and `TxOut` as the default type parameter
to maintain backward compatibility.

Add `sort_with_rng` for `TxOrdering<In, Out>` for sorting
two generic mutable slices.

This will be useful in a later commit by applying a
`TxOrdering` to the Input/Output sets of a `Selection`.